Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 int main(){
-    float x;
+    int x;
     scanf("%d",&x);
     printf("You entered: %f",x);
     return 0;
